+++ test/lein_monolith/plugin_test.clj
@@ -47,8 +47,6 @@
   (let [monolith (config/find-monolith! (read-example-project))
         subproject (-> (project/read "example/apps/app-a/project.clj")
                        (plugin/middleware monolith))]
-    (is (= '([amperity/greenlight "0.7.1"] ; This version of greenlight should come first so it'll take precedence over any other versions
-             [org.clojure/spec.alpha "0.3.218"]
-             [amperity/greenlight "0.6.0"]
-             [com.amperity/vault-clj "2.1.583"])
+    (is (= [['amperity/greenlight "0.7.1"]
+            ['org.clojure/spec.alpha "0.3.218"]]
            (:managed-dependencies subproject)))))